Refactors recreation of embeddings to solve dependency problems
