+++ controller/embedding/manager.py
@@ -5,15 +5,14 @@ from submodules.model import enums
 from util import daemon
 from . import util
 from . import connector
-from controller.misc import manager as misc
 from controller.model_provider import manager as model_manager
 from controller.project import manager as project_manager
 from submodules.model.business_objects import attribute, embedding, agreement, general
 
 
-def get_recommended_encoders() -> List[Any]:
+def get_recommended_encoders(is_managed: bool) -> List[Any]:
     recommendations = connector.request_listing_recommended_encoders()
-    if misc.check_is_managed():
+    if is_managed:
         existing_models = model_manager.get_model_provider_info()
     else:
         existing_models = []
@@ -51,32 +50,6 @@ def create_embedding(
         project_id,
         embedding_id
     )
-
-def recreate_embedding(
-    project_id: str, embedding_id: str
-) -> None:
-    old_embedding_item = embedding.get(project_id, embedding_id)
-    agreement_item = agreement.get(project_id, old_embedding_item.agreement_id)
-    new_embedding_item = embedding.create(
-        project_id,
-        old_embedding_item.attribute_id,
-        old_embedding_item.name,
-        enums.EmbeddingState.INITIALIZING.value,
-        type=old_embedding_item.type,
-        model=old_embedding_item.model,
-        platform=old_embedding_item.platform,
-        api_token=old_embedding_item.api_token,
-        with_commit=False
-    )
-    agreement_item.xfkey = new_embedding_item.id
-    general.commit()
-    connector.request_deleting_embedding(project_id, old_embedding_item.id)
-    daemon.run(
-        connector.request_embedding,
-        project_id,
-        new_embedding_item.id
-    )
-
 
 def create_embeddings_one_by_one(
     project_id: str,
